Font-awesome fixed, images, responsive changes
Font-awesome was not importing correctly, so when the build was pushed
to the server computers that did not have the font were having issues
(such as phones). Fixed with middleman font-awesome gem.

Additionally, images were a little large so that was impacting load
times in production. Image sizes were reduced again slightly.

Finally, a few responsive changes were made for small vertical
viewports to prevent the arrow down button from covering the
social icons.