New Github Pages site
Decided I'm going to go with github pages as the hosting page for my
site for the moment, and also made a transition to using hakyll for
site generation. From here, I'll probably set up CircleCI to handle
doing commits to the master branch. From then on, take a look at the
hakyll branch for active development.